requirements.txt
Flask==1.1.2
gmaps==0.9.0
gmplot==1.4.1
googlemaps==4.10.0
ipywidgets==8.0.4
mongomock==4.1.2
pymongo==3.12.0
pytest==6.2.4
python-dotenv==1.0.0
